+++ packages/coverde_cli/lib/src/features/comparison/models/models.dart
@@ -1,0 +1,1 @@
+export 'numeric_comparison.dart';