+++ packages/coverde_cli/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## UNRELEASED
 
 - **BREAKING FEAT**: full support for analyzer package v8-v10 (#263).
+- **FEAT**: add `transform` command (#276).
 
 ## 0.3.0+1